Add schematool; validate all schema refs
The logic in models/util/transform.js has been moved to the new
utils/schematool.js. The schematool also has a validator function that's
now used in the Menu and Category models so ObjectId refs are now
ensured to point to valid documents.

Also renamed the files in models/ to begin with capitals by convention.
Git only recognized the rename for the Item model so for the others,
`git log --follow <model>.js` won't show history before this commit.